prog/pspecFig.c   - Added the -normspectra, -showwedge, -normside, -intnrs and -phase options     This allows all spectra powers to be properly labeled, and scaled to avoid arbitrary numbers. The pulse longitude axis can be expressed in phase.     Note that the numbers along the intensity axis of the horizontal/vertical collapses are now 2/3 as small.     Note that the max nr along the intensity axis of the horizontal collapse is no longer missing
lib/psrio_preprocess.c
lib/psrsalsa.h
lib/application.c
  - Made change in function preprocess_fftshift(), which shouldn't affect things

lib/psrio.c
  - Improved some warnings

lib/astronomy.c
  - Increased the precision of the DM constant

lib/vonMises.c
lib/psrsalsa.h
lib/psrsalsa_typedefs.h
  - Changed von Mises definition/code from float precision to double

Patrick

+++ src/lib/psrsalsa_typedefs.h
@@ -34,7 +34,7 @@ typedef struct {
 
 typedef struct {
 
-  float centre[maxNrVonMisesComponents], concentration[maxNrVonMisesComponents], height[maxNrVonMisesComponents];
+  double centre[maxNrVonMisesComponents], concentration[maxNrVonMisesComponents], height[maxNrVonMisesComponents];
   int nrcomponents;
 }vonMises_def;
